+++ changelog.md
@@ -14,6 +14,7 @@ This changelog helps developers and users to keep track of new features, fixes a
 * test: add tests for REST data representations
 * refactor: extract issue permission verification for current user
 * fix: remove typo in configuration text
+* feat: agreement is not reached for only one vote
 
 ## 2.4.1 - 2018-04-20